+++ src/hooks/useTheme.ts
@@ -14,5 +14,4 @@ export default function useTheme() {
   })
 
   return theme
-  // return 'dark'
 }