+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+## [3.10.5](https://github.com/haydenull/logseq-plugin-agenda/compare/v3.10.4...v3.10.5) (2024-05-23)
+
+
+### Bug Fixes
+
+* **agenda2:** :bug: prevent the sync todoist button from being clicked repeatedly ([#315](https://github.com/haydenull/logseq-plugin-agenda/issues/315)) ([3183201](https://github.com/haydenull/logseq-plugin-agenda/commit/318320188c3813677165a2399a2912be661b4447)), closes [#312](https://github.com/haydenull/logseq-plugin-agenda/issues/312)
+
+## [3.10.4](https://github.com/haydenull/logseq-plugin-agenda/compare/v3.10.3...v3.10.4) (2024-05-23)
+
+
+### Bug Fixes
+
+* **kanban:** :bug: some tasks are hidden in kanban ([#314](https://github.com/haydenull/logseq-plugin-agenda/issues/314)) ([e39fe12](https://github.com/haydenull/logseq-plugin-agenda/commit/e39fe1250c9cdfca8401566b4d991c34730b7d4f))
+
+## [3.10.3](https://github.com/haydenull/logseq-plugin-agenda/compare/v3.10.2...v3.10.3) (2024-04-08)
+
+
+### Bug Fixes
+
+* **block.ts:** insert scheduled error when block has multiple lines ([#306](https://github.com/haydenull/logseq-plugin-agenda/issues/306)) ([9ede498](https://github.com/haydenull/logseq-plugin-agenda/commit/9ede49826a05048b71ece08c8cffd850bb9effde))
+
+## [3.10.2](https://github.com/haydenull/logseq-plugin-agenda/compare/v3.10.1...v3.10.2) (2024-03-29)
+
+
+### Bug Fixes
+
+* invalid duration string `DURATION:PT` ([#294](https://github.com/haydenull/logseq-plugin-agenda/issues/294)) ([70a780b](https://github.com/haydenull/logseq-plugin-agenda/commit/70a780bccf62a6fd17c252797a41aa85b8eb73ad))
+
+## [3.10.1](https://github.com/haydenull/logseq-plugin-agenda/compare/v3.10.0...v3.10.1) (2024-03-28)
+
+
+### Bug Fixes
+
+* :bug: aoto siwtch theme when logseq's theme changed ([#301](https://github.com/haydenull/logseq-plugin-agenda/issues/301)) ([e9fcb8d](https://github.com/haydenull/logseq-plugin-agenda/commit/e9fcb8d5d0f26c1ba11c1175c53f7ac0775a8ab2))
+
+# [3.10.0](https://github.com/haydenull/logseq-plugin-agenda/compare/v3.9.2...v3.10.0) (2024-03-28)
+
+
+### Features
+
+* ✨ add dark theme ([#300](https://github.com/haydenull/logseq-plugin-agenda/issues/300)) ([fa8fd6c](https://github.com/haydenull/logseq-plugin-agenda/commit/fa8fd6ceba90900a8a00aadaef7b52ba653ea93a))
+
 ## [3.9.2](https://github.com/haydenull/logseq-plugin-agenda/compare/v3.9.1...v3.9.2) (2024-01-20)